integration and unifying index archives
