+++ package.json
@@ -25,8 +25,8 @@
       "typescript": "^4.0"
    },
    "scripts": {
+      "build": "tsc",
       "start": "tsc & node build/index.js",
-      "start-postgres": "tsc & node build/src/index.js",
       "test": "echo \"Error: no test specified\" && exit 1"
    },
    "repository": {
