+++ src/entity/User.ts
@@ -9,5 +9,8 @@ export class User {
   name: string;
 
   @Column()
+  email: string;
+
+  @Column()
   password: string;
 }
