+++ tsconfig.json
@@ -14,7 +14,6 @@
 
 },
       "include": [
-         "index.ts",
          ".eslintrc.js",
          "server.js",
          "src/index.ts"
